Move HomeInnerPage union type to views.
