+++ app/views/homeEmailSentFeedback.scala.html
@@ -1,4 +1,4 @@
-@(infos: HomeController.HomeInnerPage.EmailSentFeedback)(implicit flash: Flash, request: RequestHeader)
+@(infos: views.HomeInnerPage.EmailSentFeedback)(implicit flash: Flash, request: RequestHeader)
 
 <div class="mdl-cell mdl-cell--12-col mdl-cell--middle">
     <h3 class="mdl-typography--text-uppercase mdl-typography--font-medium">
